refactor: move affine math code into math
